# docs/iotx-openapi-extracted.yaml
...
openapi: 3.0.3
info:
  title: IoTX IBS — API cho app người dùng cuối
  version: 1.0.0
  description: |
    API của máy ứng dụng IBS (NestJS) phục vụ app di động Flutter và web app.

    **Base URL** đã gồm `/v1` (xem `servers`). Mọi đường trong tài liệu này là đường TƯƠNG ĐỐI
    sau `/v1`.

    **Xác thực**: đăng nhập bằng `POST /auth/login` (hoặc `POST /auth/register`) lấy
    `accessToken` (JWT Keycloak) rồi gửi `Authorization: Bearer <accessToken>` cho mọi cửa có
    khoá. Token hết hạn (`expiresIn` giây) thì gọi `POST /auth/refresh`; refresh cũng 401 thì
    quay về màn đăng nhập.

    **Mã trạng thái**: NestJS trả `201` cho POST thành công (không phải 200); GET/PATCH/DELETE
    trả `200`.

    **Quy ước lỗi** (nguyên tắc chống dò, kế thừa pilot đã nghiệm thu):
    - `401` — thiếu/sai/hết hạn token, sai mật khẩu. Thân thường là
      `{"message":"Unauthorized","statusCode":401}` hoặc `{"message":"sai email hoặc mật khẩu"}`.
    - `404` ĐỒNG NHẤT `{"message":"not_found"}` — tenant sai, user sai tenant, thiết bị không
      tồn tại HOẶC không phải của mình (anti-IDOR), mã claim sai. Cố ý không phân biệt để không
      lộ thứ gì tồn tại. Một số lỗi nghiệp vụ (hết hạn mức, luật không hợp lệ) cũng trả `404`
      với `message` tiếng Việt — app cứ hiện `message` thẳng lên màn hình.
    - `400` — thân yêu cầu sai ở cửa nhận mạch thật / lệnh mục vật tư, `message` tiếng Việt.
    - `403` — bị chặn tạm vì thử quá nhiều lần (dò mã claim).
...

...
servers:
...
security: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      operationId: dangNhap
      summary: Đăng nhập
      description: |
        Email phải thuộc ĐÚNG tenant gửi lên (cách ly tenant tại IBS) — sai tenant trả `404`
        đồng nhất, KHÁC với sai mật khẩu (`401`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant, email, password]
              properties:
                tenant: { type: string, example: livotec }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '201':
          description: Bộ token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BoToken' }
        '401':
          description: Sai mật khẩu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loi' }
        '404': { $ref: '#/components/responses/KhongThay' }
        '429': { $ref: '#/components/responses/QuaTai' }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: lamMoiToken
      summary: Đổi refreshToken lấy bộ token mới
      description: App gọi khi accessToken hết hạn/401. Refresh cũng 401 thì quay về đăng nhập.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '201':
          description: Bộ token mới (không kèm `tenant`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BoToken' }
        '401': { $ref: '#/components/responses/KhongQuyen' }
  /me:
    get:
      tags: [auth]
      operationId: hoSo
      summary: Hồ sơ người đang đăng nhập
      security: [{ bearerJWT: [] }]
      responses:
        '200':
          description: Email, tenant và theme của người dùng
          content:
            application/json:
              schema:
                type: object
                required: [email, tenant, tenantName, theme, tbCustomerId]
                properties:
                  email: { type: string, format: email }
                  tenant: { type: string, example: livotec }
                  tenantName: { type: string, example: Livotec Home }
                  theme:
                    type: object
                    additionalProperties: true
                  tbCustomerId: { type: string, format: uuid }
        '401': { $ref: '#/components/responses/KhongQuyen' }
        '404': { $ref: '#/components/responses/KhongThay' }
  /bootstrap:
    get:
      tags: [auth]
      operationId: moApp
      summary: MỘT cửa cho lúc mở app — devices + categories + me + unread + nhãn phiên bản (B-UI.5)
      description: |
        Gộp 4 cửa app gọi lúc mở (`/devices`, `/categories`, `/me`, `/notifications` phần
        `unread`) thành một; poll giữ nhịp cũng chuyển sang đây kèm `If-None-Match` — không có
        gì đổi thì `304` thân rỗng. Các cửa lẻ VẪN GIỮ NGUYÊN cho app bản cũ.

        `phienBan` là NHÃN PHIÊN BẢN để app so với bản đang giữ, khớp thì THÔI tải cửa tương ứng:
        - `i18n` — bằng đúng `ETag` của `GET /i18n?lang&tenant` (cùng thân, cùng hàm băm)
        - `products` — nhãn đổi khi catalogue hoặc bộ chữ đổi; lệch bản đang giữ thì tải lại
          `GET /products` (KHÔNG cam kết bằng ETag của cửa đó)
        - `theme` — bằng đúng `ETag` của `GET /tenant/theme?tenant`; theme cũng nằm sẵn trong
          `me.theme` nên app đã đăng nhập không cần gọi lại cửa theme

        `ETag` của CHÍNH cửa này băm trọn thân (kể cả `lastValues`) — chỉ để poll rẻ, KHÔNG
        dùng làm nhãn phiên bản lâu dài.

        KHÔNG NHẦM với `POST /v1/admin/bootstrap` — cửa đó là Sadmin KHỞI TẠO hệ thống
        (migrate + realm + seed), khoá `x-admin-key`, không dành cho app; trùng mỗi cái tên.
      security: [{ bearerJWT: [] }]
      parameters:
        - name: lang
          in: query
          required: false
          schema: { type: string }
          description: Ngôn ngữ nối nhãn capability trong devices (mặc định en), như GET /devices
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Đủ dữ liệu dựng màn Home trong một lần hỏi
          content:
            application/json:
              schema:
                type: object
                required: [devices, categories, me, unread, phienBan]
                properties:
                  devices:
                    type: array
                    description: Y HỆT thân của GET /devices (đã gộp thiết bị thật + chia sẻ + ảo)
                    items: { $ref: '#/components/schemas/ThietBi' }
                  categories:
                    type: array
                    description: Y HỆT thân của GET /categories
                    items:
                      type: object
                      required: [kind, name]
                      properties:
                        kind: { type: string, enum: [house, room, grp] }
                        name: { type: string }
                  me:
                    type: object
                    description: Y HỆT thân của GET /me
                    required: [email, tenant, tenantName, theme, tbCustomerId]
                    properties:
                      email: { type: string, format: email }
                      tenant: { type: string, example: livotec }
                      tenantName: { type: string, example: Livotec Home }
                      theme:
                        type: object
                        additionalProperties: true
                      tbCustomerId: { type: string, format: uuid }
                  unread:
                    type: integer
                    description: Số thông báo chưa đọc — đếm như GET /notifications (trong 50 bản ghi mới nhất)
                  phienBan:
                    type: object
                    required: [i18n, products, theme]
                    properties:
                      i18n: { type: string, example: W/"jZx0f3kq1a2b3c4d" }
                      products: { type: string, example: W/"a1b2c3d4e5f6a7b8" }
                      theme: { type: string, example: W/"0f1e2d3c4b5a6978" }
        '304': { $ref: '#/components/responses/KhongDoi' }
        '401': { $ref: '#/components/responses/KhongQuyen' }
        '404': { $ref: '#/components/responses/KhongThay' }

  # ══════════════ products — catalog public ══════════════
...
  /devices:
    get:
      tags: [devices]
      operationId: dsThietBi
      summary: Mọi thiết bị của tôi + được chia sẻ cho tôi + thiết bị ảo
      description: |
        MỘT danh sách gộp: thiết bị mình claim, thiết bị người khác chia sẻ (`shared: true`,
        quyền trong `perms`) và thiết bị ảo (`virtual: true`). Mỗi hàng kèm `product`
        (capabilities + khuôn UI) và `lastValues` (số đo mới nhất) — app dựng màn Home từ đúng
        một cửa này.
      security: [{ bearerJWT: [] }]
      parameters:
        - name: lang
          in: query
          required: false
          schema: { type: string }
          description: Ngôn ngữ nối nhãn capability (mặc định en)
      responses:
        '200':
          description: Danh sách thiết bị
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ThietBi' }
        '401': { $ref: '#/components/responses/KhongQuyen' }
        '404': { $ref: '#/components/responses/KhongThay' }
...
  /devices/{id}/rpc:
    post:
      tags: [devices]
      operationId: guiLenh
      summary: Gửi lệnh điều khiển (RPC) xuống thiết bị
      description: |
        `method` phải nằm trong `capabilities[].rpc` của product — method lạ trả `404`. Với
        thiết bị ảo: `method: press` bấm nút; mọi method khác đọc `params.on` để bật/tắt trạng
        thái nhà. Header `Idempotency-Key` (tuỳ chọn, CHỈ áp cho thiết bị thật — nhánh thiết
        bị ảo trả lời trước khi tới lớp khoá): gửi lại cùng khoá thì KHÔNG chạy lệnh lần nữa,
        trả kết quả cũ kèm `replayed: true` — app nên gửi UUID mỗi lần bấm nút để an toàn khi
        mạng chập. Sau lệnh tay, tự động hoá NHƯỜNG TAY (không đụng thiết bị này ngay).
      security: [{ bearerJWT: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
          description: UUID chống lặp lệnh khi mạng chập/bấm đúp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method: { type: string, example: setPower }
                params:
                  type: object
                  additionalProperties: true
                  example: { power: true }
      responses:
        '201':
          description: 'Kết quả lệnh (`replayed: true` nếu Idempotency-Key trùng)'
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
                  replayed: { type: boolean }
                additionalProperties: true
        '401': { $ref: '#/components/responses/KhongQuyen' }
        '404': { $ref: '#/components/responses/KhongThay' }
...
  /stream:
    get:
      tags: [stream]
      operationId: luongTrangThai
      summary: Luồng SSE trạng thái thiết bị thời gian thực
      description: |
        Server-Sent Events (`Content-Type: text/event-stream`). Xác thực Bearer Y HỆT
        `/devices` — thiếu/sai token trả `401` NGAY (không mở luồng). Kết nối mở là server đẩy:

        - Sự kiện `event: trang-thai` — mỗi lần một chỉ số của thiết bị mình xem được đổi giá
          trị. `data` là JSON một dòng:
          `{"deviceId":"<id>","key":"<chỉ số>","value":<giá trị>,"ts":<epoch ms>}`
        - Nhịp giữ kết nối: dòng chú thích `: giu-ket-noi` TỐI ĐA mỗi 30 giây — client đọc quá
          30 giây không thấy gì thì coi như đứt, tự đóng và nối lại.

        Query `?thiet_bi=<id,id>` (tuỳ chọn) lọc theo danh sách id thiết bị, phân tách bằng dấu
        phẩy; vắng thì nhận mọi thiết bị mình xem được (của mình + được chia sẻ + ảo).

        Client TỰ NỐI LẠI khi đứt (backoff tuỳ client). Server KHÔNG hỗ trợ `Last-Event-ID` —
        không phát lại sự kiện đã lỡ; nối lại xong app nên gọi `GET /devices` một lần để lấy
        `lastValues` hiện tại rồi nghe tiếp.
      security: [{ bearerJWT: [] }]
      parameters:
        - name: thiet_bi
          in: query
          required: false
          schema: { type: string }
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6,v1f0c9f2a41d84f36b2f7
          description: Danh sách id thiết bị cần nghe, phân tách dấu phẩy; vắng = tất cả
      responses:
        '200':
          description: Luồng SSE mở — server đẩy sự kiện tới khi client đóng
          content:
            text/event-stream:
              schema:
                type: string
                description: Khung SSE chuẩn; xem x-sse-events
              example: |
                : giu-ket-noi

                event: trang-thai
                data: {"deviceId":"3fa85f64-5717-4562-b3fc-2c963f66afa6","key":"power","value":true,"ts":1756500000000}
        '401': { $ref: '#/components/responses/KhongQuyen' }
      x-sse-events:
        trang-thai:
          description: Một chỉ số của một thiết bị vừa đổi giá trị
          data:
            type: object
            required: [deviceId, key, value, ts]
            properties:
              deviceId: { type: string, description: Id thiết bị (thật hoặc ảo) }
              key: { type: string, description: Tên chỉ số (trùng key trong lastValues) }
              value:
                description: Giá trị mới — kiểu theo capability (boolean/số/chuỗi)
              ts: { type: integer, format: int64, description: Thời điểm — epoch mili-giây }
        heartbeat:
          description: 'Dòng chú thích `: giu-ket-noi` tối đa mỗi 30 giây, không phải event có tên'
components:
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: accessToken từ /auth/login hoặc /auth/register (JWT Keycloak)
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema: { type: string }
      description: |
        ETag nhận ở lần gọi trước (B-UI.5). Nội dung chưa đổi thì máy chủ trả `304` thân rỗng —
        app giữ nguyên bản đang cầm. ETag băm THÂN JSON (dạng yếu `W/"..."`), không dựa bộ nhớ
        tiến trình, nên nhiều bản ibs sau bộ cân bằng tải vẫn trả cùng nhãn cho cùng nội dung.
  responses:
    KhongDoi:
      description: 304 — nội dung không đổi so với ETag trong If-None-Match; thân rỗng, app dùng bản đang cầm
...
    KhongQuyen:
      description: Thiếu/sai/hết hạn token (hoặc sai mật khẩu ở cửa đăng nhập)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Loi' }
    KhongThay:
      description: |
        404 ĐỒNG NHẤT `{"message":"not_found"}` — không tồn tại HOẶC không phải của mình
        (anti-IDOR), cố ý không phân biệt. Một số lỗi nghiệp vụ (hết hạn mức, luật không hợp
        lệ…) cũng 404 với `message` tiếng Việt để app hiện thẳng.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Loi' }
    QuaTai:
      description: Hãng chạm trần hạn mức (B7) — thực tế trả 404 với message tiếng Việt, giữ mục này để client sinh sẵn nhánh xử lý
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Loi' }
...
  schemas:
    Loi:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Luôn có. `not_found` (404 đồng nhất) hoặc câu tiếng Việt hiện thẳng lên app
          example: not_found
        statusCode: { type: integer, example: 401 }
        error: { type: string, example: Bad Request }
    BoToken:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string, description: JWT — gửi 'Authorization Bearer' }
        refreshToken: { type: string, description: Đổi token mới qua /auth/refresh }
        expiresIn: { type: integer, example: 300, description: Hạn accessToken tính bằng giây }
        tenant: { type: string, description: Chỉ có ở register/login — refresh không kèm }
    Quyen:
      type: object
      description: Quyền trên thiết bị được chia sẻ — vắng trường nào coi như true
      properties:
        control: { type: boolean, description: Được điều khiển (RPC + lệnh mục) }
        create: { type: boolean, description: Được tạo (tự động hoá dùng thiết bị này…) }
        delete: { type: boolean, description: Được xoá }
    KhaNang:
      type: object
      description: Một capability của sản phẩm (hợp đồng Matter-informed, nhãn đã dịch)
      required: [key, kind, label]
      properties:
        key: { type: string, example: speed }
        kind: { type: string, enum: [onoff, level, enum, sensor, list] }
        label: { type: string, description: Nhãn đã dịch theo lang }
        rpc:
          type: string
          description: Method gửi /devices/{id}/rpc để chỉnh capability này
        min: { type: number }
        max: { type: number }
        step: { type: number }
        unit: { type: string, example: '%' }
        unitSpace: { type: boolean, description: true = có dấu cách trước đơn vị ('24 giờ') }
        values:
          type: array
          items: { type: string }
          description: kind=enum — các giá trị gửi xuống mạch
        labels:
          type: object
          additionalProperties: { type: string }
          description: kind=enum — nhãn đã dịch cho từng giá trị
      additionalProperties: true
    SanPham:
      type: object
      description: Định nghĩa một loại sản phẩm trong catalog (đã nối nhãn/đơn vị)
      required: [nameVi, category, icon, capabilities, telemetry, version]
      properties:
        nameVi: { type: string, description: Tên đã dịch theo lang (rơi về nameEn/nameVi) }
        nameEn: { type: string }
        category: { type: string, example: comfort }
        icon: { type: string, example: 🌀 }
        productGroup: { type: string, nullable: true }
        ui: { $ref: '#/components/schemas/KhuonUi' }
        capabilities:
          type: array
          items: { $ref: '#/components/schemas/KhaNang' }
        telemetry:
          type: array
          items: { type: string }
          description: Các key số đo sản phẩm đẩy lên
        version: { type: integer }
      additionalProperties: true
    KhuonUi:
      type: object
      nullable: true
      description: Màn hình riêng của sản phẩm; app chưa biết `renderer` thì RƠI VỀ màn chung, không bao giờ trắng màn
      properties:
        renderer: { type: string, example: fan }
        controls:
          type: array
          items: { type: string }
        gauges:
          type: array
          items: { type: string }
        image: { type: string, description: emoji hoặc data URI }
        slots:
          type: object
          description: Vị trí/kiểu vẽ/cỡ cho từng capability
          additionalProperties:
            type: object
            properties:
              nhom: { type: string }
              variant: { type: string }
              co: { type: integer }
      additionalProperties: true
    ThietBi:
      type: object
      description: Một hàng của GET /devices — thiết bị thật, được chia sẻ, hoặc ảo
      required:
        [id, name, type, label, house, room, grp, fav, hidden, active, createdTime, lastValues, shared, perms]
      properties:
        id: { type: string, description: UUID thiết bị thật — hoặc id bắt đầu bằng `v` cho thiết bị ảo }
        name: { type: string, description: Tên gốc (serial với mạch thật) }
        type:
          {
            type: string,
            example: fan,
            description: 'Loại sản phẩm; thiết bị ảo: virtual_external/state/button',
          }
        product:
          allOf: [{ $ref: '#/components/schemas/SanPham' }]
          nullable: true
          description: null khi sản phẩm không còn trong catalog — app vẽ màn CHUNG từ capabilities
        label: { type: string, description: Tên NGƯỜI DÙNG đặt (meta theo người xem) }
        house: { type: string, description: Rỗng = chưa phân loại }
        room: { type: string }
        grp: { type: string }
        fav: { type: boolean }
        hidden: { type: boolean }
        active: { type: boolean, description: Đang online }
        createdTime: { type: integer, format: int64, description: Epoch mili-giây }
        lastValues:
          type: object
          additionalProperties: true
          description: Số đo mới nhất theo key telemetry; capability kiểu `list` là mảng mục đã GỘP mạch + máy chủ
        shared: { type: boolean, description: true = người khác chia sẻ cho tôi }
        perms: { $ref: '#/components/schemas/Quyen' }
        virtual: { type: boolean, description: Chỉ có ở thiết bị ảo }
        virtualKind: { type: string, enum: [external, state, button] }
        place: { type: string, nullable: true, description: Thiết bị ảo nguồn ngoài — tên thành phố }
...